requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
Werkzeug==2.2.3
python-socketio==5.4.0
python-engineio==4.2.1
eventlet==0.33.0
PyJWT==2.1.0
cryptography==3.4.7
Pillow==9.5.0
requests==2.26.0
python-dotenv==0.19.0
gunicorn==20.1.0